requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
lxml==4.8.0
PyJWT==2.3.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
APScheduler==3.9.1
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
idna==3.3
lxml==4.8.0
PyJWT==2.3.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
APScheduler==3.9.1
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
idna==3.3
lxml==4.8.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
